requirements.txt
macaddress==2.0.2
pycryptodome==3.20.0
requests==2.31.0
setuptools==69.2.0
python-dotenv==1.0.1